setting torch.bfloat16

+++ src/socratic_rl_peft.py
@@ -102,7 +102,7 @@ class ActionValueFn:
         self.tokenizer.save_pretrained(path)
 
     def load(self) -> None:
-        self.model = TanhOutputModel.from_pretrained(self._base_model, num_labels=1)
+        self.model = TanhOutputModel.from_pretrained(self._base_model, num_labels=1, torch_dtype=torch.bfloat16)
         self.tokenizer = AutoTokenizer.from_pretrained(self._base_model)
         self.tokenizer.add_special_tokens({'pad_token': '[PAD]'})
         self.tokenizer.chat_template = (
